Decompose src/cli.js into per-concern CLI modules
Pure-move refactor for #386. Extract along existing seams, following the
src/commands/doctor.js per-command pattern:

- src/commands/auth.js: auth subcommand (status table + guided login);
  printBackendStatus moves with it since `auth status` owns the table and
  --list-backends imports it back. Distinct from src/auth.js (HTTP key
  resolution), now noted in both files' top comments.
- src/cli/args.js: parseArgs, mode/browser validation, option readers,
  printHelp.
- src/cli/batch.js: batch safety plan, circuit breaker, storm
  classification, batch output writing.
- src/cli/score-gate.js: applyScoreGate/extractScoreOverall; parseJsonScore
  deduped into output.js parseFirstJson (now exported, equivalent
  semantics); toFiniteScore kept local because output.js toFiniteNumber
  strips non-numeric characters before Number() and is not equivalent.
- src/cli/input.js: loadInputs and SIGINT-aware readStdin.

src/cli.js keeps main, createCancellationController, resolvePromptMode,
resolveProfileForLanguage, and the bin entry/self-invocation guard; all
existing imports from src/cli.js keep working unchanged. Function bodies
are byte-identical to main except the parseJsonScore->parseFirstJson call.
